konfigurator/forms.py
from django import forms
from django.utils.translation import ugettext_lazy as _, ugettext
from crispy_forms.bootstrap import PrependedText, Accordion, AccordionGroup, Tab, FieldWithButtons, StrictButton, PrependedAppendedText, TabHolder, FormActions
from crispy_forms import layout, bootstrap
from crispy_forms.helper import FormHelper
from crispy_forms.layout import Layout, Fieldset, ButtonHolder, Submit, MultiField, Div, Field, Button
from .models import Busy

