requirements.txt
amqp==2.1.4
appdirs==1.4.3
Babel==2.4.0
billiard==3.5.0.2
cairocffi==0.8.0
CairoSVG==2.0.2
celery==4.0.2
cffi==1.10.0
cssselect==1.0.1
defusedxml==0.5.0
Django==1.11
django-braces==1.11.0
django-crispy-forms==1.6.1
django-embed-video==1.1.0
django-localflavor==1.4.1
django-memcache-status==1.3
django-parler==1.7
django-paypal==0.3.6
django-rosetta==0.7.12
djangorestframework==3.6.2
flower==0.9.1
html5lib==0.999999999
kombu==4.0.2
lxml==3.7.3
microsofttranslator==0.8
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
polib==1.0.8
psycopg2==2.7.1
pycparser==2.17
PyJWT==1.4.2
pyparsing==2.2.0
Pyphen==0.9.4
python-memcached==1.58
python-social-auth==0.3.6
python3-memcached==1.51
python3-openid==3.1.0
pytz==2016.10
redis==2.10.3
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-core==1.2.0
sorl-thumbnail==12.3
tinycss==0.4
tornado==4.2
vine==1.1.3
WeasyPrint==0.36
webencodings==0.5
